[176] Resolve comments
